Add Dockerfile for haskell-stack projects
Stack is a nice tool, but it is not idea inside a docker image.
Stack commands do a lot of work the first time, and saves the
results for later use. Docker can't really leverage that; so
repeated builds will do some redundant downloading and compiling.

I advise using volume mounts with the script provided, so you can
rebuild less frequently. Just run the stack commands inside the
container.